+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  * Added Message Processors as a way to post-process messages before being sent out.
    Can be disabled by setting the static variable LobbyMember.MessageProcessor to null.
    Defaults to a LZFProcessor that applies LZF compression to the messages.
+ * Added the ConnectionStats struct and endpoints to monitor the stats of each connection
+   in lobbies.
 
 ## [1.0.0] - 2019-10-20
  * Initial Release